WIP: move `arch` module to root directory
