+++ arch/x86/include/arch/32/idt.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "kernel/arch/x86/gdt.h"
+#include "arch/gdt.h"
 #include <stddef.h>
 #include <stdint.h>
