+++ arch/x86/include/arch/devices/vga.h
