+++ arch/x86/src/32/gdt.c
@@ -1,5 +1,5 @@
-#include "kernel/arch/x86/gdt.h"
-#include "kernel/arch/x86/descriptor_table.h"
+#include "arch/gdt.h"
+#include "arch/descriptor_table.h"
 
 #define LIMIT_4KiB 0xfffffU
