+++ arch/x86/src/32/linker.ld
