+++ arch/x86/src/port.c
@@ -1,4 +1,4 @@
-#include "kernel/arch/x86/port.h"
+#include "arch/port.h"
 #include <stdint.h>
 
 uint8_t
